+++ crates/noseyparker-cli/tests/scan/basic/mod.rs
@@ -1,4 +1,5 @@
 use super::*;
+pub use pretty_assertions::assert_ne;
 
 #[test]
 fn scan_emptydir() {
@@ -247,3 +248,41 @@ fn report_unlimited_matches() {
         assert_cmd_snapshot!(noseyparker_success!("report", "-d", scan_env.dspath(), "--max-matches", "-1"));
     });
 }
+
+
+/// Test that the `report` command uses colors as expected when *not* running under a pty:
+///
+/// - When running with the output explicitly written to a file, colors are not used
+///
+/// - When running with with the output explicitly written to a file and `--color=always`
+///   specified, colors are used
+#[test]
+fn report_output_colors1() {
+    let scan_env = ScanEnv::new();
+    let input = scan_env.input_file_with_secret("input.txt");
+
+    let output1 = scan_env.child("findings.txt");
+    let output2 = scan_env.child("findings.colored.txt");
+
+    noseyparker_success!("scan", "-d", scan_env.dspath(), input.path())
+        .stdout(match_scan_stats("104 B", 1, 1, 1));
+
+    noseyparker_success!("report", "-d", scan_env.dspath(), "-o", output1.path());
+    noseyparker_success!("report", "-d", scan_env.dspath(), "-o", output2.path(), "--color=always");
+
+    let output1_contents = std::fs::read_to_string(output1.path()).unwrap();
+    let output2_contents = std::fs::read_to_string(output2.path()).unwrap();
+
+    assert_ne!(output1_contents, output2_contents);
+    with_settings!({
+        filters => get_report_stdout_filters(),
+    }, {
+        assert_snapshot!(output1_contents);
+    });
+    assert_eq!(&output1_contents, &console::strip_ansi_codes(&output2_contents));
+}
+
+// Test that the `report` command uses colors as expected when running under a pty:
+// - When running with the output going to stdout (default), colors are used
+// - When running with the explicitly written to a file, colors are not used
+// XXX to get a pty, look at the `pty-process` crate: https://docs.rs/pty-process/latest/pty_process/blocking/struct.Command.html